Add test for binary propagation
